services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
boto3==1.35.24
httpx==0.27.2
PyJWT==2.9.0
python-multipart==0.0.9
email-validator==2.2.0
